requirements.txt
# Core video processing
moviepy==1.0.3
manim==0.18.0

# Data and math
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
matplotlib==3.7.2

# Image processing
Pillow==10.0.0

# Audio processing
pydub==0.25.1

# Configuration
PyYAML==6.0.1

# CLI utilities
rich==13.5.2
tqdm==4.66.1

# Optional but recommended
opencv-python==4.8.0.76